+++ src/test/java/com/example/solidconnection/community/post/service/PostCommandServiceTest.java
@@ -2,6 +2,7 @@ package com.example.solidconnection.community.post.service;
 
 import com.example.solidconnection.community.board.domain.Board;
 import com.example.solidconnection.community.post.domain.Post;
+import com.example.solidconnection.community.post.domain.PostCategory;
 import com.example.solidconnection.community.post.domain.PostImage;
 import com.example.solidconnection.community.post.dto.PostCreateRequest;
 import com.example.solidconnection.community.post.dto.PostCreateResponse;
@@ -11,13 +12,11 @@ import com.example.solidconnection.community.post.dto.PostUpdateResponse;
 import com.example.solidconnection.community.post.repository.PostImageRepository;
 import com.example.solidconnection.community.post.repository.PostRepository;
 import com.example.solidconnection.custom.exception.CustomException;
+import com.example.solidconnection.s3.ImgType;
 import com.example.solidconnection.s3.S3Service;
 import com.example.solidconnection.s3.UploadedFileUrlResponse;
-import com.example.solidconnection.service.RedisService;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.support.integration.BaseIntegrationTest;
-import com.example.solidconnection.s3.ImgType;
-import com.example.solidconnection.community.post.domain.PostCategory;
 import com.example.solidconnection.util.RedisUtils;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.DisplayName;
